refactor(339): split admin-members and admin-users services post-rebase
- Extract AdminInvitationsService from admin-members (534→335+256)
- Extract AdminUsersLifecycleService from admin-users (527→398+167)
- Extract shared findUserSnapshotOrThrow utility (34 lines)
- Update controllers, modules, and tests accordingly
