+++ apps/api/src/admin/admin-invitations.controller.ts
@@ -13,7 +13,7 @@ import { Throttle } from '@nestjs/throttler'
 import { Permissions } from '../auth/decorators/permissions.decorator.js'
 import { Session } from '../auth/decorators/session.decorator.js'
 import type { AdminSession } from '../auth/types.js'
-import { AdminMembersService } from './admin-members.service.js'
+import { AdminInvitationsService } from './admin-invitations.service.js'
 import { AdminExceptionFilter } from './filters/admin-exception.filter.js'
 
 @ApiTags('Admin Invitations')
@@ -22,14 +22,14 @@ import { AdminExceptionFilter } from './filters/admin-exception.filter.js'
 @Throttle({ global: { ttl: 60_000, limit: 30 } })
 @Controller('api/admin/invitations')
 export class AdminInvitationsController {
-  constructor(private readonly adminMembersService: AdminMembersService) {}
+  constructor(private readonly adminInvitationsService: AdminInvitationsService) {}
 
   @Get()
   @Permissions('members:read')
   @ApiOperation({ summary: 'List pending invitations for the current organization' })
   @ApiResponse({ status: 200, description: 'List of pending invitations' })
   async listPendingInvitations(@Session() session: AdminSession) {
-    return this.adminMembersService.listPendingInvitations(session.session.activeOrganizationId)
+    return this.adminInvitationsService.listPendingInvitations(session.session.activeOrganizationId)
   }
 
   @Delete(':invitationId')
@@ -42,7 +42,7 @@ export class AdminInvitationsController {
     @Param('invitationId', new ParseUUIDPipe({ version: '4' })) invitationId: string,
     @Session() session: AdminSession
   ) {
-    await this.adminMembersService.revokeInvitation(
+    await this.adminInvitationsService.revokeInvitation(
       invitationId,
       session.session.activeOrganizationId,
       session.user.id
