+++ apps/api/src/admin/admin.module.ts
@@ -4,12 +4,14 @@ import { AuthModule } from '../auth/auth.module.js'
 import { AdminAuditLogsController } from './admin-audit-logs.controller.js'
 import { AdminAuditLogsService } from './admin-audit-logs.service.js'
 import { AdminInvitationsController } from './admin-invitations.controller.js'
+import { AdminInvitationsService } from './admin-invitations.service.js'
 import { AdminMembersController } from './admin-members.controller.js'
 import { AdminMembersService } from './admin-members.service.js'
 import { AdminOrganizationsController } from './admin-organizations.controller.js'
 import { AdminOrganizationsDeletionService } from './admin-organizations.deletion.js'
 import { AdminOrganizationsService } from './admin-organizations.service.js'
 import { AdminUsersController } from './admin-users.controller.js'
+import { AdminUsersLifecycleService } from './admin-users.lifecycle.js'
 import { AdminUsersService } from './admin-users.service.js'
 
 @Module({
@@ -23,14 +25,18 @@ import { AdminUsersService } from './admin-users.service.js'
   ],
   providers: [
     AdminMembersService,
+    AdminInvitationsService,
     AdminUsersService,
+    AdminUsersLifecycleService,
     AdminOrganizationsService,
     AdminOrganizationsDeletionService,
     AdminAuditLogsService,
   ],
   exports: [
     AdminMembersService,
+    AdminInvitationsService,
     AdminUsersService,
+    AdminUsersLifecycleService,
     AdminOrganizationsService,
     AdminOrganizationsDeletionService,
     AdminAuditLogsService,
